% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_indicators}
\alias{binary_indicators}
\title{Предоперационные бинарные показатели пациентов}
\format{Датафрейм с шестью полями:
\describe{
\item{Группа}{Группа, к которой относится пациент}
\item{Диабет}{Показатель №1 - диабет}
\item{ТИА}{Показатель №2 - транзиторная ишемическая атака}
\item{ХОБЛ}{Показатель №3 - хроническая обструктивная болезнь лёгких}
\item{Аритмия}{Показатель №4 - аритмия}
\item{Предшествующие вмешательства}{Показатель №5 - предшествующие вмешательства}
}}
\usage{
binary_indicators
}
\description{
Имеется выборка из N пациентов, разделенных на 3 группы, а также набор показателей.
Каждый показатель представляет собой набор бинарных значений (0 и 1). 1 - пациент имеет данный показатель, 0 - показатель у пациента отсутствует.
}
\keyword{datasets}
