% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{Категориальный признак}
\format{Датафрейм с двумя полями:
\describe{
\item{Группа}{Группа, к которой относится пациент}
\item{Категория}{Категория, к которой относится пациент}
}}
\usage{
categories
}
\description{
Имеется выборка из N пациентов, разделенных на 3 группы, и каждому пациенту соответствует какая-либо категория признака.
}
\keyword{datasets}
