% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstestplot.R
\name{cmp.categories}
\alias{cmp.categories}
\title{Построение графика сравнения категорий категориального признака в исследуемых группах пациентов}
\usage{
cmp.categories(data, group_id, group_labels, column_cat, path_pics_out,
  pic_name, p_display = "all", p_adjust_method = "BH")
}
\arguments{
\item{data}{Датафрейм с данными (Пример: statstestplot::categories)}

\item{group_id}{Число, являющееся номером столбца с группами пациентов}

\item{group_labels}{Строковый вектор, cодержащий имена групп. Эти имена будут сопоставлены в легенде графика}

\item{column_cat}{Число, являющееся номером столбца с категориями}

\item{path_pics_out}{Строка, содержащая путь, по которому будет сохранен файл с графиком}

\item{pic_name}{Строка, содержащая имя сохраняемого файла}

\item{p_display}{Решает, какие парные p-значения отображать.
Возможные значения: "s" (значимые [p <= 0.05]), "ns" (незначимые [p > 0.05]), "all" (все значения). (По умолчанию: "all").}

\item{p_adjust_method}{Решает, каким методом будут отрегулированы p-значения.
Возможные значения: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". (По умолчанию: "BH").}
}
\value{
Объект класса ggplot2
}
\description{
Имеется выборка из N пациентов, разделенных на 2 или более группы, и каждому пациенту соответствует какая-либо категория признака.
}
\details{
На графике каждая категория - это набор столбцов, их количество равно количеству групп пациентов.
Над столбцами отображается информация о количестве в группе пациентов, принадлежащих данной категории, а также процент этих людей от всех пациентов в группе.
Для столбцов в виде стрелки отображается диапазон, в котором с вероятностью 95\% находится "истинный" процент пациентов, называемый доверительным интервалом.
Для всевозможных пар групп отображается p-значение, которое используется для их сравнения.
}
\examples{
#Подключим библиотеку и загрузим датафрейм с категориями
library(statstestplot)
data(categories)

#Распечатаем его
print(categories)

#Построим график с параметрами по умолчанию
group_id <- 1
group_labels <- c("A", "B", "C")
column_cat <- 2
picture_name <- "Категории_по_умолчанию"
picture_path <- paste0(getwd(), "/")
cmp.categories(categories, group_id, group_labels, column_cat,
    picture_path, picture_name)

#Построим график, отобразив только незначимые p и отрегулировав их методом поправки Бонферрони
group_id <- 1
group_labels <- c("A", "B", "C")
column_cat <- 2
p_display <- "ns"
p_adjust_method <- "bonferroni"
picture_name <- "Категории_незначимые_p_метод_Бонферрони"
picture_path <- paste0(getwd(), "/")
cmp.categories(categories, group_id, group_labels,
    column_cat, picture_path, picture_name, p_display, p_adjust_method)
}
