% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_indicators}
\alias{binary_indicators}
\title{Бинарные показатели пациентов}
\format{Датафрейм с шестью полями:
\describe{
\item{Группа}{Группа, к которой относится пациент}
\item{Показатель_1}{Бинарный показатель №1}
\item{Показатель_2}{Бинарный показатель №2}
\item{Показатель_3}{Бинарный показатель №3}
\item{Показатель_4}{Бинарный показатель №4}
\item{Показатель_5}{Бинарный показатель №5}
}}
\usage{
binary_indicators
}
\description{
Имеется выборка из N пациентов, разделенных на 3 группы, а также набор показателей.
Каждый показатель представляет собой набор бинарных значений (0 и 1). 1 - пациент имеет данный показатель, 0 - показатель у пациента отсутствует.
}
\keyword{datasets}
