% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statstestplot.R
\name{cmp.bin.indicators}
\alias{cmp.bin.indicators}
\title{Построение графика сравнения бинарных показателей в исследуемых группах пациентов}
\usage{
cmp.bin.indicators(data, group_id, group_labels, column_list,
  path_pics_out, pic_name, p_display = "all", p_adjust_method = "BH")
}
\arguments{
\item{data}{Датафрейм с данными (Пример: statstestplot::binary_indicators)}

\item{group_id}{Число, являющееся номером столбца с группами пациентов}

\item{group_labels}{Строковый вектор, cодержащий имена групп. Эти имена будут сопоставлены в легенде графика}

\item{column_list}{Числовой вектор, содержащий номера столбцов с бинарными показателями}

\item{path_pics_out}{Строка, содержащая путь, по которому будет сохранен файл с графиком}

\item{pic_name}{Строка, содержащая имя сохраняемого файла}

\item{p_display}{Решает, какие парные p-значения отображать.
Возможные значения: "s" (значимые [p <= 0.05]), "ns" (незначимые [p > 0.05]), "all" (все значения). (По умолчанию: "all").}

\item{p_adjust_method}{Решает, каким методом будут отрегулированы p-значения.
Возможные значения: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". (По умолчанию: "BH").}
}
\value{
Объект класса ggplot2
}
\description{
Имеется выборка из N пациентов, разделенных на 2 или более группы, а также набор показателей.
Каждый показатель представляет собой набор бинарных значений (0 и 1). 1 - пациент имеет данный показатель, 0 - показатель у пациента отсутствует.
}
\details{
На графике каждый показатель - это набор столбцов, их количество равно количеству групп пациентов.
Над столбцами отображается информация о количестве в группе пациентов, имеющих данный показатель, а также процент этих людей от всех пациентов, имеющих данный показатель.
Для столбцов в виде стрелки отображается диапазон, в котором с вероятностью 95\% находится "истинный" процент пациентов, называемый доверительным интервалом.
Для всевозможных пар групп отображается p-значение, которое используется для их сравнения.
}
\examples{
#Подключим библиотеку и загрузим датафрейм с бинарными показателями
library(statstestplot)
data(binary_indicators)

#Распечатаем его
print(binary_indicators)

#Построим график с параметрами по умолчанию
group_id <- 1
group_labels <- c("A", "B", "C")
columns <- c(2, 3, 4, 5, 6)
picture_name <- "БП_все_значения_p"
picture_path <- paste0(getwd(), "/")
graphic <- cmp.bin.indicators(binary_indicators, group_id,
    group_labels, columns, picture_path, picture_name)

#Построим график с отображением только незначимых p и их регулировкой методом Бенджамини-Хохберга
group_id <- 1
group_labels <- c("A", "B", "C")
columns <- c(2, 3, 4, 5, 6)
p_display <- "ns"
p_adjust_method = "BH"
picture_name <- "БП_незначимые_p"
picture_path <- paste0(getwd(), "/")
graphic <- cmp.bin.indicators(binary_indicators, group_id,
    group_labels, columns, picture_path, picture_name, p_display, p_adjust_method)
}
